+++ src/components/index.js
@@ -14,3 +14,5 @@ export { default as DraftedBillTable } from "./DraftedBillTable";
 export { default as OutstandingBillTable } from "./OutstandingBillTable";
 export { default as CompletedBillTable } from "./CompletedBillTable";
 export { default as CancelledBillTable } from "./CancelledBIllTable";
+export { default as BillSubForm } from "./BillSubForm";
+export { default as BillFormPreview } from "./BillFormPreview";
