+++ src/containers/Billing/Deposit/DepositDetail.js
@@ -17,7 +17,7 @@ import { generateID } from "../../../utils/generateID";
 import { useReactToPrint } from "react-to-print";
 import { styled } from "@mui/material/styles";
 import { useLocation } from "react-router-dom";
-import { BackButton} from "../../../components";
+import { BackButton } from "../../../components";
 import { constants } from "../../../utils/constants";
 
 const StyledTypography = styled(Typography)(({ theme }) => ({
@@ -232,7 +232,7 @@ const CategoryDetail = () => {
                   {details?.amount} MMK
                 </StyledTypography>
               </Box>
-            </Box> 
+            </Box>
           </Box>
         </Box>
         {/* <Box sx={{ flexDirection: "column", padding: "20px 10px" }}>
