+++ src/recoil/billForm/atom.js
@@ -1,0 +1,12 @@
+import { atom } from "recoil";
+
+const billFormAtom = atom({
+  key: "billFormAtom",
+  default: {
+    currentPatient: null,
+    totalDeposit: 0,
+    billItems: [],
+  },
+});
+
+export default billFormAtom;
