+++ src/recoil/billForm/index.js
@@ -1,0 +1,8 @@
+import atom from "./atom";
+import withCurrentPatient from "./withCurrentPatient";
+import withBillItems from "./withBilItems";
+import withTotalDeposit from "./withTotalDeposit";
+
+export { withCurrentPatient, withBillItems, withTotalDeposit };
+
+export default atom;
